modified entities from JPA to mongoDB, removed boilerplate code
