--- src/main/java/com/vehiclerental/vehicle_rental_app/Constants/CommonConstants.java
@@ -1,4 +1,0 @@
-package com.vehiclerental.vehicle_rental_app.Constants;
-
-public class CommonConstants {
-}
