+++ src/main/java/com/vehiclerental/vehicle_rental_app/constants/CommonConstants.java
@@ -1,0 +1,4 @@
+package com.vehiclerental.vehicle_rental_app.constants;
+
+public class CommonConstants {
+}
