+++ src/main/java/com/vehiclerental/vehicle_rental_app/enums/ReservationStatus.java
@@ -1,7 +1,6 @@
 package com.vehiclerental.vehicle_rental_app.enums;
 
 public enum ReservationStatus {
-    ON_HOLD,
     BOOKED,
     CANCELLED
 }
