+++ src/main/java/com/vehiclerental/vehicle_rental_app/enums/VehicleStatus.java
@@ -2,6 +2,7 @@ package com.vehiclerental.vehicle_rental_app.enums;
 
 public enum VehicleStatus {
     AVAILABLE,
+    RESERVED,
     RENTED,
     MAINTENANCE
 }
