+++ src/main/java/com/vehiclerental/vehicle_rental_app/repositories/CustomerRepository.java
@@ -1,0 +1,9 @@
+package com.vehiclerental.vehicle_rental_app.repositories;
+
+import com.vehiclerental.vehicle_rental_app.entities.CustomerEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CustomerRepository extends MongoRepository<CustomerEntity, String> {
+}
