+++ src/main/resources/application.properties
@@ -3,3 +3,7 @@ server.port=8080
 
 spring.security.oauth2.resourceserver.jwt.issuer-uri=http://localhost:8081/realms/vehiclerental
 spring.security.oauth2.resourceserver.jwt.jwk-set-uri=http://localhost:8081/realms/vehiclerental/protocol/openid-connect/certs
+
+spring.data.mongodb.database=vra
+spring.data.mongodb.uri=mongodb://localhost:27017
+
